Triangle!
The triangle location is a bit messed up but the viewport may be the issue.
Added shader code to draw the triangle.
Added window autocentre because I was sick of having to move the window to my screen centre each time.
Learnt that when uploading data to the buffers, you need to heap allocate the stuff, pretty interesting.

Signed-off-by: Alistair <alistair@alistairbell.xyz>